pyproject.toml
[project]
name = "sft-dataset-maker"
version = "0.1.0"
description = "监督微调数据集制作工具"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.23",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[tool.uv]

[dependency-groups]
dev = []
